requirements.txt
openai
mcp
anyio
python-dotenv
pydantic
google-api-python-client
google-auth
google-auth-oauthlib
